I have changed around the structure of the xspec.fit package, and started work on the annular spectrum set fitting code for an absorbed plasma model
