+++ xga/xspec/fit/__init__.py
@@ -1,8 +1,9 @@
 #  This code is a part of XMM: Generate and Analyse (XGA), a module designed for the XMM Cluster Survey (XCS).
 #  Last modified by David J Turner (david.turner@sussex.ac.uk) 21/01/2021, 11:45. Copyright (c) David J Turner
 
-from .fit import single_temp_apec, power_law, single_temp_apec_profile
-from .run import execute_cmd, xspec_call
+# Perhaps this one will be removed
+from .profile import *
+from .single import *
